Remove indirection between arguments and environment
Previously, the arguments of a FlatOp indexed into the list of arguments of a cluster. These arguments then indexed into the 'actual' environment. With this change, the arguments of a FlatOp directly index into the 'actual' environment. By removing this indirection we can also directly refer to the individual sizes of the different dimensions of the different arrays (which may be useful if we include more information on indices in clusters)