If two variables have same constant value, substitute them
Previously, when we would detect that two variables had the same constant value, matchVars', as used in detectCopy for instance, would use the InfoEnv to check if the variables have the same value. This worked, but made things more complicated. When we didn't have a InfoEnv, we couldn't accurately detect if two variables are the same (especially to check if two arrays have the same size).

With this change, when a variable is introduced with a constant value, we check if a previous variable was introduced with the same constant value. If that is the case, we substitute the later variable with the first variable. This makes it easier to check if arrays have the same size, as we can now just check if the variables are the same.

As an alternative, we could consider to allow either a variable or a constant on the places where we now accept variables. This is for instance what LLVM does with Operands, which can either be a constant or a variable. However, this would require many changes in the compiler.